refactor: изменен метод getEpicStatus, Enum вынесен в отдельный файл, изменены типы переменных и др.

+++ src/Epic.java
@@ -1,6 +1,6 @@
 public class Epic extends Task {
 
-    public Epic(Integer id, String name, String details, Status status) {
+    public Epic(long id, String name, String details, Status status) {
         super(id, name, details, status);
     }
